# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My Barber API",
    "version": "1.0.0",
    "description": "Barber booking API built with Node.js and TypeScript.\n\n**Schemas:** `components.schemas` mirrors TypeScript types in `models/` (e.g. `BookingStatus`, `BookingContract` ↔ `models/booking.ts`; barber approval ↔ `models/barber.ts`). Enumerations are OpenAPI `enum`s; request/response bodies use `$ref` where noted.",
    "contact": {
      "name": "API Support",
      "email": "support@my-barber.uz"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.my-barber.uz",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.my-barber.uz",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Barber/client JWT from login endpoints, or a Firebase Auth ID token with custom claim admin:true / role:admin for Admin routes. Dev/bootstrap: JWT whose user `id` is in ADMIN_UID_ALLOWLIST."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          },
          "details": {
            "type": "array",
            "description": "Optional express-validator errors (ValidationError[] style).",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
      "BarberApprovalStatus": {
        "type": "string",
        "description": "Barber moderation state (models/barber.ts BarberApprovalStatus).",
        "enum": [
          "pending",
          "approved",
          "rejected",
          "suspended"
        ]
      },
...
      "BarberResponse": {
        "type": "object",
        "description": "Public barber profile (models/barber.ts BarberResponse; no password field).",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "additionalProperties": true
          },
          "birthDate": {
            "type": "string"
          },
          "workingHours": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "approvalStatus": {
            "$ref": "#/components/schemas/BarberApprovalStatus"
          },
          "approvalMessage": {
            "type": "string",
            "nullable": true
          },
          "ratingAverage": {
            "type": "number",
            "nullable": true
          },
          "ratingCount": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "AdminBarberListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "barbers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BarberResponse"
                }
              },
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              }
            }
          }
        }
      },
      "AdminApprovalPatchBody": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "suspended"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "CmsBannerDoc": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "imageRef": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CmsBannerDocInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "imageRef": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "AdminStatsSummary": {
        "type": "object",
        "properties": {
          "barbersByApprovalStatus": {
            "type": "object",
            "properties": {
              "pending": {
                "type": "integer"
              },
              "approved": {
                "type": "integer"
              },
              "rejected": {
                "type": "integer"
              },
              "suspended": {
                "type": "integer"
              },
              "unknown": {
                "type": "integer"
              }
            }
          },
          "bookingsLast7Days": {
            "type": "integer"
          },
          "bookingsLast30Days": {
            "type": "integer"
          }
        }
      }
    }
  },
...
  "paths": {
    "/admin/barbers": {
      "get": {
        "summary": "List barbers (moderation)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "approvalStatus",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "suspended"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated barbers (no passwords)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBarberListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — not admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/barbers/{barberId}": {
      "get": {
        "summary": "Get a single barber (admin)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Barber profile (no password)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "barber": {
                          "$ref": "#/components/schemas/BarberResponse"
                        },
                        "services": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Barber not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/barbers/{barberId}/approval": {
      "patch": {
        "summary": "Update barber approval status",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminApprovalPatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated barber profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BarberResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition or validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Barber not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/content/banner": {
      "get": {
        "summary": "Get CMS banner doc",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Banner config (or empty defaults)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CmsBannerDoc"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Upsert CMS banner doc",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CmsBannerDocInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved banner",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CmsBannerDoc"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/stats/summary": {
      "get": {
        "summary": "Aggregated stats (barbers by approval, recent bookings)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Summary counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminStatsSummary"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/barbers": {
      "get": {
        "summary": "Get all barbers with pagination",
        "description": "Same behavior as GET /barber. The barber router is mounted at both `/barber` and `/barbers`.",
        "tags": [
          "Barber"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Number of items per page (max 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of barbers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "barbers": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}